# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API for people management following hexagonal architecture principles",
        "title": "Pessoas API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@pessoasapi.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/persons": {
            "get": {
                "description": "Returns a paginated list of persons with sorting options",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persons"
                ],
                "summary": "List persons with pagination",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "default": 10,
                        "description": "Items per page",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "id",
                            "name",
                            "cpf",
                            "email",
                            "created_at",
                            "updated_at"
                        ],
                        "type": "string",
                        "default": "id",
                        "description": "Field to sort by",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "asc",
                            "desc"
                        ],
                        "type": "string",
                        "default": "desc",
                        "description": "Sort direction",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/contract.PaginatedResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/contract.PersonResponseDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Creates a new person in the system with the provided data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persons"
                ],
                "summary": "Create a new person",
                "parameters": [
                    {
                        "description": "Person data to be created",
                        "name": "person",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/contract.NewPersonDTO"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/contract.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "Business validation error",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/persons/cpf/{cpf}": {
            "get": {
                "description": "Returns person data based on the provided Brazilian CPF",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Persons"
                ],
                "summary": "Find person by CPF",
                "parameters": [
                    {
                        "type": "string",
                        "example": "111.444.777-35",
                        "description": "Person's CPF (with or without formatting)",
                        "name": "cpf",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/contract.PersonResponseDTO"
                        }
                    },
                    "400": {
                        "description": "CPF not provided",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Person not found",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/contract.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "contract.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error code",
                    "type": "string",
                    "example": "validation_error"
                },
                "message": {
                    "description": "Descriptive error message",
                    "type": "string",
                    "example": "Invalid CPF"
                }
            }
        },
        "contract.NewPersonDTO": {
            "type": "object",
            "required": [
                "birth_date",
                "cpf",
                "email",
                "name",
                "phone"
            ],
            "properties": {
                "birth_date": {
                    "description": "Date of birth",
                    "type": "string",
                    "example": "1990-01-15T00:00:00Z"
                },
                "cpf": {
                    "description": "Brazilian CPF (can be formatted or digits only)",
                    "type": "string",
                    "example": "111.444.777-35"
                },
                "email": {
                    "description": "Valid email address",
                    "type": "string",
                    "example": "joao.silva@email.com"
                },
                "name": {
                    "description": "Person's full name",
                    "type": "string",
                    "example": "João Silva"
                },
                "phone": {
                    "description": "Phone number (10 or 11 digits)",
                    "type": "string",
                    "example": "81912345678"
                }
            }
        },
        "contract.PaginatedResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "List of items in current page",
                    "type": "array",
                    "items": {
                        "type": "object"
                    }
                },
                "page": {
                    "description": "Current page number",
                    "type": "integer",
                    "example": 1
                },
                "page_size": {
                    "description": "Number of items per page",
                    "type": "integer",
                    "example": 10
                },
                "total_items": {
                    "description": "Total number of items in database",
                    "type": "integer",
                    "example": 100
                },
                "total_pages": {
                    "description": "Total number of pages available",
                    "type": "integer",
                    "example": 10
                }
            }
        },
        "contract.PersonResponseDTO": {
            "type": "object",
            "properties": {
                "birth_date": {
                    "description": "Date of birth",
                    "type": "string",
                    "example": "1990-01-15T00:00:00Z"
                },
                "cpf": {
                    "description": "Brazilian CPF (digits only)",
                    "type": "string",
                    "example": "11144477735"
                },
                "created_at": {
                    "description": "Record creation timestamp",
                    "type": "string",
                    "example": "2024-01-01T10:00:00Z"
                },
                "email": {
                    "description": "Email address",
                    "type": "string",
                    "example": "joao.silva@email.com"
                },
                "id": {
                    "description": "Unique person ID",
                    "type": "integer",
                    "example": 1
                },
                "name": {
                    "description": "Full name",
                    "type": "string",
                    "example": "João Silva"
                },
                "phone": {
                    "description": "Phone number (digits only)",
                    "type": "string",
                    "example": "81912345678"
                },
                "updated_at": {
                    "description": "Last update timestamp",
                    "type": "string",
                    "example": "2024-01-01T10:00:00Z"
                }
            }
        },
        "contract.SuccessResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "ID of created resource",
                    "type": "integer",
                    "example": 1
                },
                "message": {
                    "description": "Success message",
                    "type": "string",
                    "example": "Person created successfully"
                }
            }
        }
    },
    "tags": [
        {
            "description": "Endpoints for API health check",
            "name": "Health"
        },
        {
            "description": "CRUD operations for person management",
            "name": "Persons"
        }
    ]
}